requirements.txt
scapy>=2.5.0
pandas>=2.0.0
requests>=2.28.0
python-dateutil>=2.8.2
tqdm>=4.66.0
pyyaml>=6.0